fix(records): move clearRecordPipelineClosedOutSync into the record write door
c7d8d249fb added clearRecordPipelineClosedOutSync to src/lib/pan-dir/records.ts
(plural), where its call to the raw writer writeIssueRecordSync violates the
single-write-door tenet and turned main red:
record-writers-locked.test.ts allowlists only record.ts and record-update.ts.

Moves the function verbatim into the door module record-update.ts, which already
imports ensureIssueRecordSync/writeIssueRecordSync/queueIssueRecordCommit
legitimately, and repoints the reopen.ts import. records.ts drops the three raw
writer imports the reopen fix added; nothing else there used them.

record-update.ts rather than record.ts (where the markRecordPipelineClosedOutSync
setter lives) because record.ts sits exactly at the 1000-line file-size ceiling —
adding the clearer there fails scripts/lint-file-size.sh.

Reopen behavior is unchanged.

Fixes PAN-3038
